Create translation demonology in the revival epoch (re-naissance / or rebirth)
Who else is just learning that the "re-naissance" was the "re-birth," as in, when the story of Christ was changed to be only about a father who is jealous and vengeful, leaving out the divine mother (feminine) for all non-catholics. Is that why we're the only species that doesn't hold women in esteem today, publicly, though privately every man in this country knows if mama aint happy, ain't nobody happy? Let's be smarter than retarded and under arrest? What say you?